move stuff into util/
