+++ src/fmt_md_inlines.rs
@@ -2,7 +2,7 @@ use crate::footnote_transform::FootnoteTransformer;
 use crate::link_transform::{LinkLabel, LinkTransform, LinkTransformation, LinkTransformer};
 use crate::md_elem::elem::*;
 use crate::md_elem::*;
-use crate::output::{Output, SimpleWrite};
+use crate::util::output::{Output, SimpleWrite};
 use serde::Serialize;
 use std::borrow::Cow;
 use std::cmp::max;
@@ -453,8 +453,8 @@ impl TitleQuote {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::output::{Output, OutputOpts};
-    use crate::utils_for_test::*;
+    use crate::util::output::{Output, OutputOpts};
+    use crate::util::utils_for_test::*;
 
     mod title_quoting {
         use super::*;
