+++ src/md_elem/tree.rs
@@ -836,7 +836,7 @@ impl Lookups {
 mod tests {
     use super::*;
     use crate::md_elem::tree_test_utils::*;
-    use crate::utils_for_test::*;
+    use crate::util::utils_for_test::*;
 
     impl MdContext {
         pub fn empty() -> Self {
